Move AppInfo into its own file
